services/worker-pdf/requirements.txt
# FastAPI + Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Celery + Redis
celery[redis]==5.4.0
redis==5.2.1

# Data Analysis
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5

# Visualization
plotly==5.24.1
kaleido==0.2.1

# PDF Export
weasyprint==63.1

# AI / LLM
langchain==0.3.14
langchain-openai==0.3.1
langchain-groq==0.2.3
langchain-google-genai==2.0.7
langchain-ollama==0.2.0
langgraph==0.2.60
langgraph-checkpoint-redis>=0.1.0

qdrant-client[fastembed]>=1.11.0
pypdf==5.1.0
pdf2image==1.17.0

# Advanced RAG (ColPali/MUVERA)
colpali-engine==0.3.4
transformers>=4.45.0
torch>=2.4.0
einops>=0.8.0
pillow>=10.4.0

# Encryption
cryptography==44.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Rate Limiting
slowapi==0.1.9

# Logging
structlog==24.4.0
